use LayersByType *everywhere* and fix lots of bugs in PTLayer etc.
