+++ axon/logging.go
@@ -6,7 +6,6 @@ package axon
 
 import (
 	"github.com/emer/emergent/elog"
-	"github.com/emer/emergent/emer"
 	"github.com/emer/emergent/estats"
 	"github.com/emer/emergent/etime"
 	"github.com/emer/etable/agg"
@@ -37,8 +36,8 @@ func LogTestErrors(lg *elog.Logs) {
 
 // PCAStats computes PCA statistics on recorded hidden activation patterns
 // from Analyze, Trial log data
-func PCAStats(net emer.Network, lg *elog.Logs, stats *estats.Stats) {
-	stats.PCAStats(lg.IdxView(etime.Analyze, etime.Trial), "ActM", net.LayersByClass("HiddenLayer", "TargetLayer", "CTLayer"))
+func PCAStats(net *Network, lg *elog.Logs, stats *estats.Stats) {
+	stats.PCAStats(lg.IdxView(etime.Analyze, etime.Trial), "ActM", net.LayersByType(SuperLayer, TargetLayer, CTLayer))
 }
 
 //////////////////////////////////////////////////////////////////////////////
@@ -48,7 +47,7 @@ func PCAStats(net emer.Network, lg *elog.Logs, stats *estats.Stats) {
 // across two given time levels, in higher to lower order, e.g., Epoch, Trial
 // These are useful for tuning and diagnosing the behavior of the network.
 func LogAddDiagnosticItems(lg *elog.Logs, net *Network, times ...etime.Times) {
-	layers := net.LayersByClass("Hidden", "CT", "Target")
+	layers := net.LayersByType(SuperLayer, CTLayer, TargetLayer)
 	for _, lnm := range layers {
 		clnm := lnm
 		lg.AddItem(&elog.Item{
@@ -101,7 +100,7 @@ func LogAddDiagnosticItems(lg *elog.Logs, net *Network, times ...etime.Times) {
 	}
 
 	// input layer average activity -- important for tuning
-	layers = net.LayersByClass("InputLayer")
+	layers = net.LayersByType(InputLayer)
 	for _, lnm := range layers {
 		clnm := lnm
 		lg.AddItem(&elog.Item{
@@ -121,7 +120,7 @@ func LogAddDiagnosticItems(lg *elog.Logs, net *Network, times ...etime.Times) {
 // across 3 given time levels, in higher to lower order, e.g., Run, Epoch, Trial
 // These are useful for diagnosing the behavior of the network.
 func LogAddPCAItems(lg *elog.Logs, net *Network, times ...etime.Times) {
-	layers := net.LayersByClass("HiddenLayer", "TargetLayer", "CTLayer")
+	layers := net.LayersByType(SuperLayer, TargetLayer, CTLayer)
 	for _, lnm := range layers {
 		clnm := lnm
 		cly := net.LayerByName(clnm)
@@ -184,7 +183,7 @@ func LogAddPCAItems(lg *elog.Logs, net *Network, times ...etime.Times) {
 // for given mode and time (e.g., Test, Cycle)
 // These are useful for monitoring layer activity during testing.
 func LogAddLayerGeActAvgItems(lg *elog.Logs, net *Network, mode etime.Modes, etm etime.Times) {
-	layers := net.LayersByClass("HiddenLayer", "TargetLayer")
+	layers := net.LayersByType(SuperLayer, TargetLayer)
 	for _, lnm := range layers {
 		clnm := lnm
 		lg.AddItem(&elog.Item{
@@ -212,7 +211,7 @@ func LogAddLayerGeActAvgItems(lg *elog.Logs, net *Network, mode etime.Modes, etm
 // across two given time levels, in higher to lower order, e.g., Epoch, Trial
 // These are useful for tuning and diagnosing the behavior of the network.
 func LogAddExtraDiagnosticItems(lg *elog.Logs, net *Network, times ...etime.Times) {
-	layers := net.LayersByClass("HiddenLayer", "CTLayer", "TargetLayer")
+	layers := net.LayersByType(SuperLayer, CTLayer, TargetLayer)
 	for _, lnm := range layers {
 		clnm := lnm
 		lg.AddItem(&elog.Item{
@@ -265,7 +264,7 @@ func LogAddExtraDiagnosticItems(lg *elog.Logs, net *Network, times ...etime.Time
 // These were useful for the development of the Ca-based "trace" learning rule
 // that directly uses NMDA and VGCC-like spiking Ca
 func LogAddCaLrnDiagnosticItems(lg *elog.Logs, net *Network, times ...etime.Times) {
-	layers := net.LayersByClass("HiddenLayer", "TargetLayer")
+	layers := net.LayersByType(SuperLayer, TargetLayer)
 	for _, lnm := range layers {
 		clnm := lnm
 		// ss.Logs.AddItem(&elog.Item{
@@ -438,7 +437,7 @@ func LogAddCaLrnDiagnosticItems(lg *elog.Logs, net *Network, times ...etime.Time
 // aggregated across three time scales, ordered from higher to lower,
 // e.g., Run, Epoch, Trial.
 func LogAddPulvCorSimItems(lg *elog.Logs, net *Network, times ...etime.Times) {
-	layers := net.LayersByClass("Pulvinar")
+	layers := net.LayersByType(PulvinarLayer)
 	for _, lnm := range layers {
 		clnm := lnm
 		lg.AddItem(&elog.Item{
