+++ examples/boa/boa.go
@@ -546,7 +546,7 @@ func (ss *Sim) TakeAction(net *axon.Network) {
 	genAct := ev.ActGen()
 	genActNm := ev.Acts[genAct]
 	ss.Stats.SetString("NetAction", anm)
-	ss.Stats.SetString("GenAction", genActNm)
+	ss.Stats.SetString("InstinctAction", genActNm)
 	if netAct == genAct {
 		ss.Stats.SetFloat("ActMatch", 1)
 	} else {
@@ -620,7 +620,7 @@ func (ss *Sim) ApplyInputs() {
 		ss.Sim.CortexDriving = erand.BoolProb(float64(ss.Sim.PctCortex), -1)
 		net.InitActs() // this is still essential even with fully functioning decay below:
 		// todo: need a more selective US gating mechanism!
-		net.DecayStateByClass(&ss.Context, 1, 1, "HiddenLayer", "PTLayer", "CTLayer", "VThalLayer")
+		net.DecayStateByType(&ss.Context, 1, 1, axon.SuperLayer, axon.PTMaintLayer, axon.CTLayer, axon.VThalLayer)
 		ev.RenderLocalist("Gate", 0)
 	}
 
@@ -685,11 +685,11 @@ func (ss *Sim) InitStats() {
 	ss.Stats.SetFloat("WrongCSGate", 0)
 	ss.Stats.SetFloat("Rew", 0)
 	ss.Stats.SetString("NetAction", "")
-	ss.Stats.SetString("GenAction", "")
+	ss.Stats.SetString("InstinctAction", "")
 	ss.Stats.SetString("ActAction", "")
 	ss.Stats.SetFloat("ActMatch", 0)
 	ss.Stats.SetFloat("AllGood", 0)
-	lays := ss.Net.LayersByClass("PTLayer")
+	lays := ss.Net.LayersByType(axon.PTMaintLayer)
 	for _, lnm := range lays {
 		ss.Stats.SetFloat("Maint"+lnm, 0)
 		ss.Stats.SetFloat("MaintFail"+lnm, 0)
@@ -711,7 +711,7 @@ func (ss *Sim) StatCounters() {
 	// ss.Stats.SetFloat32("ACCNeg", ss.Sim.ACCNeg)
 	// trlnm := fmt.Sprintf("pos: %g, neg: %g", ss.Sim.ACCPos, ss.Sim.ACCNeg)
 	ss.Stats.SetString("TrialName", "trl")
-	ss.ViewUpdt.Text = ss.Stats.Print([]string{"Run", "Epoch", "Trial", "Cycle", "NetAction", "GenAction", "ActAction", "ActMatch", "Gated", "Should", "Rew"})
+	ss.ViewUpdt.Text = ss.Stats.Print([]string{"Run", "Epoch", "Trial", "Cycle", "NetAction", "InstinctAction", "ActAction", "ActMatch", "Gated", "Should", "Rew"})
 }
 
 // TrialStats computes the trial-level statistics.
@@ -804,7 +804,7 @@ func (ss *Sim) MaintStats() {
 	isCons := ev.LastAct == ev.ActMap["Consume"]
 	actThr := float32(0.05) // 0.1 too high
 	net := ss.Net
-	lays := net.LayersByClass("PTLayer")
+	lays := net.LayersByType(axon.PTMaintLayer)
 	hasMaint := false
 	for _, lnm := range lays {
 		mnm := "Maint" + lnm
@@ -853,7 +853,7 @@ func (ss *Sim) ConfigLogs() {
 	ss.Logs.AddStatStringItem(etime.AllModes, etime.AllTimes, "RunName")
 	// ss.Logs.AddStatStringItem(etime.AllModes, etime.Trial, "TrialName")
 	ss.Logs.AddStatFloatNoAggItem(etime.AllModes, etime.AllTimes, "PctCortex")
-	ss.Logs.AddStatStringItem(etime.AllModes, etime.Trial, "NetAction", "GenAction", "ActAction")
+	ss.Logs.AddStatStringItem(etime.AllModes, etime.Trial, "NetAction", "InstinctAction", "ActAction")
 	// ss.Logs.AddStatFloatNoAggItem(etime.AllModes, etime.AllTimes, "ACCPos")
 	// ss.Logs.AddStatFloatNoAggItem(etime.AllModes, etime.AllTimes, "ACCNeg")
 
@@ -902,7 +902,7 @@ func (ss *Sim) ConfigLogItems() {
 	ss.Logs.AddStatAggItem("GatedPostCS", "GatedPostCS", etime.Run, etime.Epoch, etime.Trial)
 	ss.Logs.AddStatAggItem("WrongCSGate", "WrongCSGate", etime.Run, etime.Epoch, etime.Trial)
 
-	lays := ss.Net.LayersByClass("PTLayer")
+	lays := ss.Net.LayersByType(axon.PTMaintLayer)
 	for _, lnm := range lays {
 		nm := "Maint" + lnm
 		ss.Logs.AddStatAggItem(nm, nm, etime.Run, etime.Epoch, etime.Trial)
@@ -934,7 +934,7 @@ func (ss *Sim) ConfigLogItems() {
 		Write: elog.WriteMap{
 			etime.Scope(etime.Train, etime.Epoch): func(ctx *elog.Context) {
 				ix := ctx.Logs.IdxView(ctx.Mode, etime.Trial)
-				spl := split.GroupBy(ix, []string{"GenAction"})
+				spl := split.GroupBy(ix, []string{"InstinctAction"})
 				split.AggTry(spl, "ActMatch", agg.AggMean)
 				ags := spl.AggsToTable(etable.ColNameOnly)
 				ss.Logs.MiscTables["ActCor"] = ags
@@ -950,7 +950,7 @@ func (ss *Sim) ConfigLogItems() {
 			Write: elog.WriteMap{
 				etime.Scope(etime.Train, etime.Epoch): func(ctx *elog.Context) {
 					ags := ss.Logs.MiscTables["ActCor"]
-					rw := ags.RowsByString("GenAction", anm, etable.Equals, etable.UseCase)
+					rw := ags.RowsByString("InstinctAction", anm, etable.Equals, etable.UseCase)
 					if len(rw) > 0 {
 						ctx.SetFloat64(ags.CellFloat("ActMatch", rw[0]))
 					}
